chore(directories): restructure directories
