+++ src/app/bootstrap/setup-currencies.ts
@@ -1,6 +1,6 @@
 import ICurrencyRepo from '../../domain/currency/repos/currency.repo';
 import generateUUID from '../../shared/utils/uuid-generator';
-import ILogger from '../contracts/infra-services/logger.contract';
+import ILogger from '../contracts/infra/logger.contract';
 import { SYSTEM_CURRENCIES } from './data/currencies';
 
 export default async function bootstrapCurrencies(
@@ -21,7 +21,6 @@ export default async function bootstrapCurrencies(
 
     logger.info('Currencies bootstrapped successfully');
   } catch (error) {
-    console.log();
     logger.error('Failed to bootstrap currencies', error);
   }
 }
