--- src/app/contracts/dto/errors.dto.ts
@@ -1,7 +1,0 @@
-import { IApiValidationError } from '../../../shared/value-objects/error';
-
-export interface IApiError {
-  message: string;
-  validationErrors?: IApiValidationError[];
-  cause?: any;
-}
