+++ src/app/contracts/infra/auth-service.contract.ts
