+++ src/app/contracts/infra/logger.contract.ts
