+++ src/app/usecases/currency/__specs__/get-currencies.usecase.spec.ts
@@ -1,6 +1,6 @@
 import getCurrenciesUseCase from '../get-currencies.usecase';
 import { MockCurrencyRepo } from '../../../../infra/db/repos/__mocks__/currency.repo.impl.mock';
-import MockRequestContext from '../../../contracts/storage/__mocks__/request-context.mock';
+import MockRequestContext from '../../../contracts/app/__mocks__/request-context.mock';
 import { ICurrency } from '../../../../domain/currency/types/currency.types';
 import IRequestContext from '../../../../shared/types/request-context.types';
