--- src/app/usecases/ledger/create-bank-account.usecase.ts
@@ -1,3 +1,0 @@
-/**
- * See sequence diagram here ../__docs__/create-bank-account.usecase.mermaid
- */
