+++ src/infra/observability/__mocks__/reporter.mock.ts
@@ -1,4 +1,4 @@
-import IReporter from '../../../app/contracts/infra-services/reporter.contract';
+import IReporter from '../../../app/contracts/infra/reporter.contract';
 
 export const MockReporter: jest.Mocked<IReporter> = {
   reportAppError: jest.fn(),
