+++ src/infra/observability/reporter.ts
@@ -1,7 +1,7 @@
 import Sentry from '@sentry/node';
 import { NODE_ENV, SENTRY_DSN } from '../config/vars.config';
 import logger from './logger';
-import IReporter from '../../app/contracts/infra-services/reporter.contract';
+import IReporter from '../../app/contracts/infra/reporter.contract';
 
 Sentry.init({ dsn: SENTRY_DSN, sendDefaultPii: true, environment: NODE_ENV });
