+++ src/interface/http/middlewares/is-optional-authenticated-user.middleware.ts
@@ -1,7 +1,7 @@
 import { RequestHandler } from 'express';
 import httpErrorHandler from '../handlers/error.handler';
-import ILogger from '../../../app/contracts/infra-services/logger.contract';
-import IReporter from '../../../app/contracts/infra-services/reporter.contract';
+import ILogger from '../../../app/contracts/infra/logger.contract';
+import IReporter from '../../../app/contracts/infra/reporter.contract';
 
 export default function isOptionalAuthenticatedUserMiddleware(
   logger: ILogger,
